move Main into App, use hooks to track page views in App.
